pyproject.toml
[tool.poetry]
name = "neurons"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
langchain = "^0.2.2"
langchain-community = "^0.2.3"
langchain-openai = "^0.1.8"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
